requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
attrs==25.3.0
bcrypt==4.2.0
certifi==2025.7.9
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
motor==3.7.1
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.15.5
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
